fix up data use validation tests
