+++ tests/ops/api/v1/endpoints/test_privacy_notice_endpoints.py
@@ -5,9 +5,15 @@ from typing import Any, Dict
 from uuid import uuid4
 
 import pytest
+from fideslang import DataUse as DataUse
+from starlette.exceptions import HTTPException
 from starlette.testclient import TestClient
 
+from fides.api.ctl.sql_models import DataUse as sql_DataUse
 from fides.api.ops.api.v1 import scope_registry as scopes
+from fides.api.ops.api.v1.endpoints.privacy_notice_endpoints import (
+    validate_notice_data_uses,
+)
 from fides.api.ops.api.v1.urn_registry import (
     PRIVACY_NOTICE,
     PRIVACY_NOTICE_DETAIL,
@@ -20,6 +26,79 @@ from fides.api.ops.models.privacy_notice import (
     PrivacyNoticeHistory,
     PrivacyNoticeRegion,
 )
+from fides.api.ops.schemas.privacy_notice import PrivacyNoticeCreation
+
+
+class TestValidateDataUses:
+    @pytest.fixture(scope="function")
+    def privacy_notice_request(self):
+        return PrivacyNoticeCreation(
+            name="sample privacy notice",
+            regions=[PrivacyNoticeRegion.us_ca],
+            consent_mechanism=ConsentMechanism.opt_in,
+            data_uses=["placeholder"],
+            enforcement_level=EnforcementLevel.system_wide,
+        )
+
+    @pytest.fixture(scope="function")
+    def custom_data_use(self, db):
+        return sql_DataUse.create(
+            db=db,
+            data=DataUse(
+                fides_key="new_data_use",
+                organization_fides_key="default_organization",
+                name="New data use",
+                description="A test data use",
+                parent_key=None,
+                is_default=True,
+            ).dict(),
+        )
+
+    @pytest.mark.usefixtures("load_default_data_uses")
+    def test_validate_data_uses_invalid(
+        self, db, privacy_notice_request: PrivacyNoticeCreation
+    ):
+        privacy_notice_request.data_uses = ["invalid_data_use"]
+        with pytest.raises(HTTPException):
+            validate_notice_data_uses([privacy_notice_request], db)
+
+        privacy_notice_request.data_uses = ["advertising", "invalid_data_use"]
+        with pytest.raises(HTTPException):
+            validate_notice_data_uses([privacy_notice_request], db)
+
+        privacy_notice_request.data_uses = [
+            "advertising",
+            "advertising.invalid_data_use",
+        ]
+        with pytest.raises(HTTPException):
+            validate_notice_data_uses([privacy_notice_request], db)
+
+    @pytest.mark.usefixtures("load_default_data_uses")
+    def test_validate_data_uses_default_taxonomy(
+        self, db, privacy_notice_request: PrivacyNoticeCreation
+    ):
+        privacy_notice_request.data_uses = ["advertising"]
+        validate_notice_data_uses([privacy_notice_request], db)
+        privacy_notice_request.data_uses = ["advertising", "provide"]
+        validate_notice_data_uses([privacy_notice_request], db)
+        privacy_notice_request.data_uses = ["advertising", "provide", "provide.service"]
+        validate_notice_data_uses([privacy_notice_request], db)
+
+    @pytest.mark.usefixtures("load_default_data_uses")
+    def test_validate_data_uses_custom_uses(
+        self,
+        db,
+        privacy_notice_request: PrivacyNoticeCreation,
+        custom_data_use: sql_DataUse,
+    ):
+        """
+        Ensure custom data uses added to the DB are considered valid
+        """
+
+        privacy_notice_request.data_uses = [custom_data_use.fides_key]
+        validate_notice_data_uses([privacy_notice_request], db)
+        privacy_notice_request.data_uses = ["advertising", custom_data_use.fides_key]
+        validate_notice_data_uses([privacy_notice_request], db)
 
 
 class TestGetPrivacyNotices:
